requirements.txt
streamlit==1.44.1
openai==1.75.0
pillow
